+++ lib/init.php
@@ -30,9 +30,7 @@ if (!defined('__DIR__')) { define('__DIR__', dirname(__FILE__)); }
 // Define helper constants
 $get_theme_name = explode('/themes/', get_template_directory());
 
-define('WP_BASE',                   wp_base_dir());
 define('THEME_NAME',                next($get_theme_name));
-define('RELATIVE_PLUGIN_PATH',      str_replace(site_url() . '/', '', plugins_url()));
-define('FULL_RELATIVE_PLUGIN_PATH', WP_BASE . '/' . RELATIVE_PLUGIN_PATH);
-define('RELATIVE_CONTENT_PATH',     str_replace(site_url() . '/', '', content_url()));
+define('RELATIVE_PLUGIN_PATH',      str_replace(home_url() . '/', '', plugins_url()));
+define('RELATIVE_CONTENT_PATH',     str_replace(home_url() . '/', '', content_url()));
 define('THEME_PATH',                RELATIVE_CONTENT_PATH . '/themes/' . THEME_NAME);