+++ lib/utils.php
@@ -85,43 +85,6 @@ function roots_title() {
   }
 }
 
-/**
- * Show an admin notice if .htaccess isn't writable
- */
-function roots_htaccess_writable() {
-  if (!is_writable(get_home_path() . '.htaccess')) {
-    if (current_user_can('administrator')) {
-      add_action('admin_notices', create_function('', "echo '<div class=\"error\"><p>" . sprintf(__('Please make sure your <a href="%s">.htaccess</a> file is writable ', 'roots'), admin_url('options-permalink.php')) . "</p></div>';"));
-    }
-  }
-}
-if (current_theme_supports('h5bp-htaccess')) {
-  add_action('admin_init', 'roots_htaccess_writable');
-}
-
-/**
- * Return WordPress subdirectory if applicable
- */
-function wp_base_dir() {
-  preg_match('!(https?://[^/|"]+)([^"]+)?!', site_url(), $matches);
-  if (count($matches) === 3) {
-    return end($matches);
-  } else {
-    return '';
-  }
-}
-
-/**
- * Opposite of built in WP functions for trailing slashes
- */
-function leadingslashit($string) {
-  return '/' . unleadingslashit($string);
-}
-
-function unleadingslashit($string) {
-  return ltrim($string, '/');
-}
-
 function add_filters($tags, $function) {
   foreach($tags as $tag) {
     add_filter($tag, $function);