+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-### HEAD
+### 6.4.0: May 1st, 2013
+* Fix Theme Activation page issues
+* Fix issues with root relative URLs and rewrites on non-standard setups
+* Make sure rewrites are added to .htaccess immediately after activation
+* Move HTML5 Boilerplate's .htaccess to a [plugin](https://github.com/retlehs/wp-h5bp-htaccess)
+* Rename page-custom.php to template-custom.php
 * Don't warn about unwritable htaccess if that option is disabled
 * Add missing collapse class for top navbar
 * Add comment template